Move negative keyword toggle into dedicated component for easier loading state
